support EDD & fix memory access, volumes
 - fix address mapping when ramdisk accesses bootrom area
 - separate volume controls

+++ src/memory.h
@@ -14,8 +14,8 @@ private:
     uint8_t bytes[TOTAL_MEMORY];
     bool mode_stack;
     bool mode_map;
-    uint8_t page_map;
-    uint8_t page_stack;
+    uint32_t page_map;
+    uint32_t page_stack;
 
     std::vector<uint8_t> bootbytes;
